cc: float values in the value_t runtime (AIPL --xinu-jit)
value_t gains a float variant: a boxed double tagged with bit 60 (our
identity-mapped pointers fit in 48 bits, so the tag is free and masked
off before deref).  v_floatlit(bits)/v_floatval(double) box it; v_add/
v_sub/v_mul/v_div/v_lt/v_le/v_eq/v_print now promote int+float -> float
and render floats with 6 fractional digits.  Ints stay immediate.

The runtime needs FP, which the kernel otherwise forbids
(-mgeneral-regs-only): cc.c is now built without that flag (Makefile
per-file override) and mmu_init enables EL1 FP access via CPACR.  The
on-device compiler stays integer-only; the lexer accumulates numeric
literals as unsigned so 64-bit float bit patterns survive.

QEMU-verified: a Rotor.spin AIPL program accumulates 0.5 per step
(3 -> 1.500000) and "angle = " + 2.0 -> "angle = 2.000000".

+++ cc/parse.c
@@ -94,9 +94,12 @@ token_t *cc_lex(const char *src)
         }
 
         if (is_digit(*p)) {
-            long v = 0;
-            while (is_digit(*p)) { v = v * 10 + (*p - '0'); p++; }
-            token_t *t = new_tok(TK_NUM); t->val = v;
+            /* unsigned accumulation so 64-bit constants (e.g. IEEE-754
+             * float bit patterns from --xinu-jit) wrap to the right
+             * two's-complement value instead of signed-overflow UB. */
+            unsigned long v = 0;
+            while (is_digit(*p)) { v = v * 10UL + (unsigned long)(*p - '0'); p++; }
+            token_t *t = new_tok(TK_NUM); t->val = (long)v;
             cur = cur->next = t;
             continue;
         }
